+++ tnxjee-service/src/main/java/org/truenewx/tnxjee/service/exception/AuthenticationFailureException.java
@@ -1,9 +1,11 @@
 package org.truenewx.tnxjee.service.exception;
 
+import org.springframework.security.core.AuthenticationException;
+
 /**
  * 鉴权失败异常
  */
-public class AuthenticationFailureException extends BusinessException {
+public class AuthenticationFailureException extends AuthenticationException {
 
     private static final long serialVersionUID = -7826129157270679401L;
